+++ src/App.jsx
@@ -2,10 +2,11 @@ import { Routes, Route } from 'react-router-dom';
 import Navbar from './components/navbar';
 import Program from './components/program';
 import Home from './components/home';
-import Levels from './components/levels'
+import Levels from './components/Levels'
 import Nosotros from './components/nosotros';
 import Contact from './components/contact';
 import './App.css';
+import Instalaciones from './components/Instalaciones';
 
 function App() {
   return (
@@ -16,6 +17,7 @@ function App() {
         <Route path='nosotros' element={<Nosotros />} />
         <Route path='programa' element={<Program />} />
         <Route path='niveles' element={<Levels />} />
+        <Route path='instalaciones' element={<Instalaciones />} />
         <Route path='contacto' element={<Contact />} />
         <Route path='*' element={<Home />} />
       </Routes>
